+++ truss/patch/signature.py
@@ -2,8 +2,8 @@ from pathlib import Path
 from typing import List, Optional
 
 from truss.constants import CONFIG_FILE
+from truss.patch.custom_types import TrussSignature
 from truss.patch.dir_signature import directory_content_signature
-from truss.patch.types import TrussSignature
 from truss.truss_config import TrussConfig
