+++ truss/patch/truss_dir_patch_applier.py
@@ -2,6 +2,16 @@ import logging
 from pathlib import Path
 from typing import List
 
+from truss.templates.control.control.helpers.custom_types import (
+    Action,
+    ConfigPatch,
+    EnvVarPatch,
+    ExternalDataPatch,
+    ModelCodePatch,
+    Patch,
+    PythonRequirementPatch,
+    SystemPackagePatch,
+)
 from truss.templates.control.control.helpers.errors import UnsupportedPatch
 from truss.templates.control.control.helpers.truss_patch.model_code_patch_applier import (
     apply_code_patch,
@@ -12,16 +22,6 @@ from truss.templates.control.control.helpers.truss_patch.requirement_name_identi
 )
 from truss.templates.control.control.helpers.truss_patch.system_packages import (
     system_packages_set,
-)
-from truss.templates.control.control.helpers.types import (
-    Action,
-    ConfigPatch,
-    EnvVarPatch,
-    ExternalDataPatch,
-    ModelCodePatch,
-    Patch,
-    PythonRequirementPatch,
-    SystemPackagePatch,
 )
 from truss.truss_config import TrussConfig
