+++ truss/remote/baseten/api.py
@@ -3,7 +3,7 @@ from enum import Enum
 from typing import Any, List, Optional
 
 import requests
-from truss.remote.baseten import types as b10_types
+from truss.remote.baseten import custom_types as b10_types
 from truss.remote.baseten.auth import ApiKey, AuthService
 from truss.remote.baseten.error import ApiError
 from truss.remote.baseten.utils.transfer import base64_encoded_json_str
