+++ truss/templates/control/control/helpers/truss_patch/model_code_patch_applier.py
@@ -4,10 +4,10 @@ from pathlib import Path
 
 # TODO(pankaj) In desparate need of refactoring into separate library
 try:
-    from helpers.types import Action, Patch
+    from helpers.custom_types import Action, Patch
 except ModuleNotFoundError as exc:
     logging.debug(f"Importing helpers from truss core, caused by: {exc}")
-    from truss.templates.control.control.helpers.types import Action, Patch
+    from truss.templates.control.control.helpers.custom_types import Action, Patch
 
 
 def apply_code_patch(
