+++ truss/templates/trtllm-audio/packages/whisper_trt/batching.py
@@ -8,7 +8,7 @@ import torch
 from async_batcher.batcher import AsyncBatcher
 from torch import Tensor
 
-from whisper_trt.types import DEFAULT_NUM_BEAMS, BatchWhisperItem
+from whisper_trt.custom_types import DEFAULT_NUM_BEAMS, BatchWhisperItem
 
 FIXED_TEXT_PRFIX = "<|startoftranscript|><|en|><|transcribe|><|0.00|>"
