+++ truss/tests/patch/test_types.py
@@ -1,5 +1,5 @@
+from truss.patch.custom_types import TrussSignature
 from truss.patch.signature import calc_truss_signature
-from truss.patch.types import TrussSignature
 
 
 def test_truss_signature_type(custom_model_truss_dir):
